+++ core/networking/src/main/kotlin/com/kelsos/mbrc/core/networking/protocol/base/Protocol.kt
@@ -70,6 +70,7 @@ sealed class Protocol(val context: String) {
   object LibraryBrowseAlbums : Protocol(LIBRARY_BROWSE_ALBUMS)
   object LibraryBrowseTracks : Protocol(LIBRARY_BROWSE_TRACKS)
   object LibraryCover : Protocol(LIBRARY_COVER)
+  object LibraryPlayAll : Protocol(LIBRARY_PLAY_ALL)
   object PlayerOutput : Protocol(PLAYER_OUTPUT)
   object PlayerOutputSwitch : Protocol(PLAYER_OUTPUT_SWITCH)
   object VerifyConnection : Protocol(VERIFY_CONNECTION)
@@ -140,6 +141,7 @@ sealed class Protocol(val context: String) {
     const val PLAYER_OUTPUT_SWITCH = "playeroutputswitch"
 
     const val LIBRARY_COVER = "libraryalbumcover"
+    const val LIBRARY_PLAY_ALL = "libraryplayall"
 
     /**
      * Toggle action in protocol. This should be send to the functions with multiple states
@@ -196,7 +198,8 @@ sealed class Protocol(val context: String) {
         COMMAND_UNAVAILABLE to CommandUnavailable,
         PLAYER_OUTPUT to PlayerOutput,
         PLAYER_OUTPUT_SWITCH to PlayerOutputSwitch,
-        LIBRARY_COVER to LibraryCover
+        LIBRARY_COVER to LibraryCover,
+        LIBRARY_PLAY_ALL to LibraryPlayAll
       )
     }